+++ lua/github-theme/extra/init.lua
@@ -3,17 +3,23 @@ package.path = "./lua/?/init.lua;./lua/?.lua"
 local configModule = require("github-theme.config")
 
 local function write(str, fileName)
-  print("[write] extra/" .. fileName)
-  local file = io.open("extras/" .. fileName, "w")
+  print("[write]" .. fileName)
+  local file = io.open(fileName, "w")
   file:write(str)
   file:close()
 end
 
-local extras = {kitty = "conf", alacritty = "yml", iterm = "itermcolors", konsole="colorscheme"}
+local extras = {
+  kitty = "conf",
+  alacritty = "yml",
+  iterm = "itermcolors",
+  konsole = "colorscheme"
+}
 for _, style in ipairs({"dark", "dimmed", "light"}) do
   configModule.themeStyle = style
   for extra, ext in pairs(extras) do
     local plugin = require("github-theme.extra." .. extra)
-    write(plugin[extra](configModule), extra .. "_github_" .. style .. "." .. ext)
+    local fileName = "extras/" .. extra .. "/" .. style .. "." .. ext
+    write(plugin[extra](configModule), fileName)
   end
 end